requirements.txt
boto3==1.13.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
pandas==1.0.3
pymysql==0.9.3
scikit-learn==0.23.1
geopy==1.22.0
numpy==1.18.5
s3fs==0.4.2
pytest==5.4.1